feat(config): OpenRouter instances default to Kimi K3 xhigh, Opus 5 judge, GLM 5.2 triage
An OpenRouter key is gateway-specific and reaches every vendor, so an
instance holding one has said something deliberate about routing. Make
that the whole engine posture rather than only a default model:

- OpenRouter moves to the head of the provider preference table, and its
  default model becomes moonshotai/kimi-k3 instead of openrouter/auto
  (a per-turn router breaks prompt-cache prefixes and makes no two turns
  comparable).
- provider_engine_baseline() supplies the posture that table row cannot:
  xhigh + thinking for the default agent, anthropic/claude-opus-5 for the
  judge, z-ai/glm-5.2 for triage. It composes UNDER the settings scope
  chain, so any field a user set stays theirs.
- Keyed on the provider resolution actually picked, not on which keys
  exist, so --model anthropic/... stays an Anthropic session.

Two fixes this depends on:
- map_openrouter_effort no longer collapses xhigh/max to high. The
  gateway now documents the full ladder and normalizes it onto the routed
  model; the collapse made a pinned xhigh silently mean high.
- AgentEngineConfig::overlay dropped headless_scope_bypass, so a
  higher-precedence scope lost that field. Latent until something layered
  underneath the user's file; load-bearing now.

The provider table moves to config/providers.rs: config.rs was already
over the 1500-line ratchet at HEAD, and the gate's own message prescribes
a submodule over a baseline entry.
